:truck: refactor(api): move usecase to configuration
The `complementary-certification` context is not really a good contexte,
we need to move code to real contexte
