--- api/src/certification/complementary-certification/domain/usecases/index.js
@@ -1,42 +1,0 @@
-// eslint-disable import/no-restricted-paths
-import { injectDependencies } from '../../../../shared/infrastructure/utils/dependency-injection.js';
-import * as complementaryCertificationForTargetProfileAttachmentRepository from '../../../complementary-certification/infrastructure/repositories/complementary-certification-for-target-profile-attachment-repository.js';
-import * as organizationRepository from '../../../complementary-certification/infrastructure/repositories/organization-repository.js';
-import * as complementaryCertificationRepository from '../../../configuration/infrastructure/repositories/complementary-certification-repository.js';
-import { mailService } from '../../../shared/domain/services/mail-service.js';
-import * as targetProfileHistoryRepository from '../../../shared/infrastructure/repositories/target-profile-history-repository.js';
-import * as complementaryCertificationBadgesRepository from '../../infrastructure/repositories/complementary-certification-badge-repository.js';
-
-/**
- *
- * Using {@link https://jsdoc.app/tags-type "Closure Compiler's syntax"} to document injected dependencies
- *
- * @typedef {complementaryCertificationBadgesRepository} ComplementaryCertificationBadgesRepository
- * @typedef {complementaryCertificationForTargetProfileAttachmentRepository} ComplementaryCertificationForTargetProfileAttachmentRepository
- * @typedef {targetProfileHistoryRepository} TargetProfileHistoryRepository
- * @typedef {organizationRepository} OrganizationRepository
- * @typedef {mailService} MailService
- **/
-const dependencies = {
-  complementaryCertificationBadgesRepository,
-  complementaryCertificationForTargetProfileAttachmentRepository,
-  targetProfileHistoryRepository,
-  organizationRepository,
-  complementaryCertificationRepository,
-  mailService,
-};
-
-import { attachBadges } from './attach-badges.js';
-import { sendTargetProfileNotifications } from './send-target-profile-notifications.js';
-
-const usecasesWithoutInjectedDependencies = {
-  attachBadges,
-  sendTargetProfileNotifications,
-};
-
-const usecases = injectDependencies(usecasesWithoutInjectedDependencies, dependencies);
-
-/**
- * @typedef {dependencies} dependencies
- */
-export { usecases };
