+++ api/src/certification/configuration/application/attach-target-profile-controller.js
@@ -1,4 +1,3 @@
-import { usecases as complementaryCertificationUsecases } from '../../complementary-certification/domain/usecases/index.js';
 import { usecases } from '../domain/usecases/index.js';
 import * as complementaryCertificationBadgeSerializer from '../infrastructure/serializers/complementary-certification-badge-serializer.js';
 
@@ -11,7 +10,7 @@ const attachTargetProfile = async function (request, h, dependencies = { complem
     complementaryCertificationId,
   });
 
-  await complementaryCertificationUsecases.attachBadges({
+  await usecases.attachBadges({
     userId,
     complementaryCertification,
     targetProfileIdToDetach: targetProfileId,
@@ -19,7 +18,7 @@ const attachTargetProfile = async function (request, h, dependencies = { complem
   });
 
   if (!!targetProfileId && notifyOrganizations) {
-    await complementaryCertificationUsecases.sendTargetProfileNotifications({
+    await usecases.sendTargetProfileNotifications({
       targetProfileIdToDetach: targetProfileId,
       complementaryCertification,
     });
