+++ api/src/certification/configuration/domain/usecases/attach-badges.js
