+++ api/src/certification/configuration/domain/usecases/index.js
@@ -50,6 +50,7 @@ const dependencies = {
   versionsRepository,
 };
 
+import { attachBadges } from './attach-badges.js';
 import { calibrateFrameworkVersion } from './calibrate-framework-version.js';
 import { catchingUpCandidateReconciliation } from './catching-up-candidate-reconciliation.js';
 import { createCertificationVersion } from './create-certification-version.js';
@@ -64,25 +65,28 @@ import { getFrameworkHistory } from './get-framework-history.js';
 import { getScoBlockedAccessDates } from './get-sco-blocked-access-dates.js';
 import { importScoWhitelist } from './import-sco-whitelist.js';
 import { searchAttachableTargetProfiles } from './search-attachable-target-profiles.js';
+import { sendTargetProfileNotifications } from './send-target-profile-notifications.js';
 import { updateCertificationVersion } from './update-certification-version.js';
 import { updateScoBlockedAccessDate } from './update-sco-blocked-access-date.js';
 
 const usecasesWithoutInjectedDependencies = {
+  attachBadges,
   calibrateFrameworkVersion,
   catchingUpCandidateReconciliation,
   createCertificationVersion,
   exportScoWhitelist,
   findCertificationFrameworks,
   findComplementaryCertifications,
-  getCurrentFrameworkVersion,
   getActiveVersionByScope,
   getComplementaryCertificationForTargetProfileAttachmentRepository,
   getComplementaryCertificationTargetProfileHistory,
+  getCurrentFrameworkVersion,
   getFrameworkHistory,
+  getScoBlockedAccessDates,
   importScoWhitelist,
   searchAttachableTargetProfiles,
+  sendTargetProfileNotifications,
   updateCertificationVersion,
-  getScoBlockedAccessDates,
   updateScoBlockedAccessDate,
 };
