+++ api/tests/certification/configuration/unit/application/attach-target-profile-controller_test.js
@@ -1,55 +1,9 @@
-import { usecases as complementaryCertificationUsecases } from '../../../../../src/certification/complementary-certification/domain/usecases/index.js';
 import { attachTargetProfileController } from '../../../../../src/certification/configuration/application/attach-target-profile-controller.js';
 import { usecases } from '../../../../../src/certification/configuration/domain/usecases/index.js';
 import { expect, hFake, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Application | Certification | ComplementaryCertification | attach-target-profile-controller', function () {
   describe('#attachTargetProfile', function () {
-    context('when there is no notification', function () {
-      it('should call the usecase and serialize the response', async function () {
-        // given
-        const request = {
-          payload: Symbol('args'),
-          auth: { credentials: { userId: 99 } },
-          params: { complementaryCertificationId: 101 },
-        };
-
-        const complementaryCertification = Symbol('certification');
-        const complementaryCertificationBadges = Symbol('complementaryCertificationBadges');
-
-        sinon.stub(usecases, 'getComplementaryCertificationForTargetProfileAttachmentRepository');
-        sinon.stub(complementaryCertificationUsecases, 'attachBadges');
-        sinon.stub(complementaryCertificationUsecases, 'sendTargetProfileNotifications');
-        const complementaryCertificationBadgeSerializerStub = {
-          deserialize: sinon.stub(),
-        };
-
-        usecases.getComplementaryCertificationForTargetProfileAttachmentRepository.resolves(complementaryCertification);
-        complementaryCertificationBadgeSerializerStub.deserialize.withArgs(request.payload).returns({
-          targetProfileId: 1,
-          notifyOrganizations: false,
-          complementaryCertificationBadges,
-        });
-
-        const dependencies = {
-          complementaryCertificationBadgeSerializer: complementaryCertificationBadgeSerializerStub,
-        };
-
-        // when
-        const result = await attachTargetProfileController.attachTargetProfile(request, hFake, dependencies);
-
-        // then
-        expect(result.statusCode).to.equal(204);
-        expect(complementaryCertificationUsecases.attachBadges).to.have.been.calledWith({
-          userId: 99,
-          complementaryCertification,
-          targetProfileIdToDetach: 1,
-          complementaryCertificationBadgesToAttachDTO: complementaryCertificationBadges,
-        });
-        expect(complementaryCertificationUsecases.sendTargetProfileNotifications).not.to.have.been.called;
-      });
-    });
-
     context('when there are notifications', function () {
       context('when there is no previous target profile', function () {
         it('should not call the notification service', async function () {
@@ -64,8 +18,8 @@ describe('Unit | Application | Certification | ComplementaryCertification | atta
           const complementaryCertificationBadges = Symbol('complementaryCertificationBadges');
 
           sinon.stub(usecases, 'getComplementaryCertificationForTargetProfileAttachmentRepository');
-          sinon.stub(complementaryCertificationUsecases, 'attachBadges');
-          sinon.stub(complementaryCertificationUsecases, 'sendTargetProfileNotifications');
+          sinon.stub(usecases, 'attachBadges');
+          sinon.stub(usecases, 'sendTargetProfileNotifications');
           const complementaryCertificationBadgeSerializerStub = {
             deserialize: sinon.stub(),
           };
@@ -87,7 +41,7 @@ describe('Unit | Application | Certification | ComplementaryCertification | atta
           await attachTargetProfileController.attachTargetProfile(request, hFake, dependencies);
 
           // then
-          expect(complementaryCertificationUsecases.sendTargetProfileNotifications).not.to.have.been.called;
+          expect(usecases.sendTargetProfileNotifications).not.to.have.been.called;
         });
       });
 
@@ -103,8 +57,8 @@ describe('Unit | Application | Certification | ComplementaryCertification | atta
         const complementaryCertificationBadges = Symbol('complementaryCertificationBadges');
 
         sinon.stub(usecases, 'getComplementaryCertificationForTargetProfileAttachmentRepository');
-        sinon.stub(complementaryCertificationUsecases, 'attachBadges');
-        sinon.stub(complementaryCertificationUsecases, 'sendTargetProfileNotifications');
+        sinon.stub(usecases, 'attachBadges');
+        sinon.stub(usecases, 'sendTargetProfileNotifications');
         const complementaryCertificationBadgeSerializerStub = {
           deserialize: sinon.stub(),
         };
@@ -125,13 +79,13 @@ describe('Unit | Application | Certification | ComplementaryCertification | atta
 
         // then
         expect(result.statusCode).to.equal(204);
-        expect(complementaryCertificationUsecases.attachBadges).to.have.been.calledWith({
+        expect(usecases.attachBadges).to.have.been.calledWith({
           userId: 99,
           complementaryCertification,
           targetProfileIdToDetach: 1,
           complementaryCertificationBadgesToAttachDTO: complementaryCertificationBadges,
         });
-        expect(complementaryCertificationUsecases.sendTargetProfileNotifications).to.have.been.calledWith({
+        expect(usecases.sendTargetProfileNotifications).to.have.been.calledWith({
           targetProfileIdToDetach: 1,
           complementaryCertification,
         });
