+++ api/tests/certification/configuration/unit/domain/usecases/attach-badges_test.js
@@ -1,5 +1,5 @@
 import { InvalidBadgeLevelError } from '../../../../../../src/certification/complementary-certification/domain/errors.js';
-import { attachBadges } from '../../../../../../src/certification/complementary-certification/domain/usecases/attach-badges.js';
+import { attachBadges } from '../../../../../../src/certification/configuration/domain/usecases/attach-badges.js';
 import { DomainTransaction } from '../../../../../../src/shared/domain/DomainTransaction.js';
 import { MissingAttributesError, NotFoundError } from '../../../../../../src/shared/domain/errors.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../../../test-helper.js';
